requirements.txt
clickhouse-connect>=0.6.0
python-dotenv>=1.0.0
psutil>=5.9.0
requests>=2.28.0
polars>=0.20.0
psycopg2-binary>=2.9.9
redis>=5.0.0